creating a nan_value property for the data that can be used to complete nan values in pythons.
raising error in interger data if a non integer is passed.
